Move main processing functions out into separate files
