+++ internal/handler/parserfilter.go
@@ -1,0 +1,7 @@
+package handler
+
+type parserfilter interface {
+
+
+}
+
